// src/protos/expansion.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.6.1
//   protoc               v5.27.1
// source: expansion.proto

/* eslint-disable */
import { Observable } from "rxjs";

export const protobufPackage = "expansion";

/** 文件块信息 */
export interface FileChunk {
  /** 文件块内容 */
  content: Uint8Array;
  /** 文件名 */
  filename: string;
  /** 块索引 */
  chunkIndex: number;
  /** 总块数 */
  totalChunks: number;
  /** 文件唯一标识 */
  fileId: string;
}

/** 上传响应 */
export interface UploadResponse {
  /** 是否成功 */
  success: boolean;
  /** 响应消息 */
  message: string;
  /** 文件URL */
  fileUrl: string;
  /** 文件ID */
  fileId: string;
}

export interface ExpansionService {
  /** 流式上传大文件 - 将文件分割成多个小块进行传输 */
  uploadZipFileStream(request: Observable<FileChunk>): Promise<UploadResponse>;
}